Event.java OneToMany 삭제 및 삭제에 따른 delete메소드 추가, requestDTO 전송 완료 log 주석으로 추가.

+++ src/main/java/com/f2z/gach/Event/Controller/AdminEventController.java
@@ -67,7 +67,13 @@ public class AdminEventController {
         model.addAttribute("eventDto", AdminEventRequestDTO
                 .toEventRequestDTO(eventRepository.findByEventId(eventId),
                 eventLocationRepository.findAllByEvent_EventId(eventId)));
-        // 확인해보니, 여기서 순회로 인한 StackOverflow 발생함.
+
+        // AdminEventRequestDTO(event=Event(eventId=1, eventName=가천대학교 축구리그, eventLink=https://www.instagram.com/p/C4zwuIOv4ga/?utm_source=ig_web_copy_link&igsh=MzRlODBiNWFlZA==, eventStartDate=2024-04-11, eventEndDate=2024-05-29, eventInfo=새로운 에너지가 충만한 2024년, 우리의 열정이 폭발하는 이곳 [2024 가천대학교 축구리그: G-LEAGUE], eventImageName=SoccerEvent.png, eventImagePath=http://ceprj.gachon.ac.kr:60002/image/SoccerEvent.png),
+        // locations=[EventLocation(eventLocationId=1, eventName=가천대학교 축구 리그, eventPlaceName=경기장, eventAltitude=75.28107668, eventLatitude=37.455084, eventLongitude=127.135115, event=Event(eventId=1, eventName=가천대학교 축구리그, eventLink=https://www.instagram.com/p/C4zwuIOv4ga/?utm_source=ig_web_copy_link&igsh=MzRlODBiNWFlZA==, eventStartDate=2024-04-11, eventEndDate=2024-05-29, eventInfo=새로운 에너지가 충만한 2024년, 우리의 열정이 폭발하는 이곳 [2024 가천대학교 축구리그: G-LEAGUE], eventImageName=SoccerEvent.png, eventImagePath=http://ceprj.gachon.ac.kr:60002/image/SoccerEvent.png)),
+        // EventLocation(eventLocationId=2, eventName=가천대학교 축구 리그, eventPlaceName=유니폼 배부 장소, eventAltitude=83.5, eventLatitude=37.454586, eventLongitude=127.13504, event=Event(eventId=1, eventName=가천대학교 축구리그, eventLink=https://www.instagram.com/p/C4zwuIOv4ga/?utm_source=ig_web_copy_link&igsh=MzRlODBiNWFlZA==, eventStartDate=2024-04-11, eventEndDate=2024-05-29, eventInfo=새로운 에너지가 충만한 2024년, 우리의 열정이 폭발하는 이곳 [2024 가천대학교 축구리그: G-LEAGUE], eventImageName=SoccerEvent.png, eventImagePath=http://ceprj.gachon.ac.kr:60002/image/SoccerEvent.png)),
+        // EventLocation(eventLocationId=3, eventName=가천대학교 축구 리그, eventPlaceName=음료수 프로모션, eventAltitude=74.0, eventLatitude=37.454792, eventLongitude=127.135404, event=Event(eventId=1, eventName=가천대학교 축구리그, eventLink=https://www.instagram.com/p/C4zwuIOv4ga/?utm_source=ig_web_copy_link&igsh=MzRlODBiNWFlZA==, eventStartDate=2024-04-11, eventEndDate=2024-05-29, eventInfo=새로운 에너지가 충만한 2024년, 우리의 열정이 폭발하는 이곳 [2024 가천대학교 축구리그: G-LEAGUE], eventImageName=SoccerEvent.png, eventImagePath=http://ceprj.gachon.ac.kr:60002/image/SoccerEvent.png))],
+        // file=null)
+
         // FIXME 여기서 객체가 전달되어야, location이 전송되는지 확인 가능함.
         // 전송하는 객체의 log.info가 정상적으로 나오면 푸시해줘. html 안 만져도 돼ㅠㅠ 고맙다.
         return "event/event-detail";
@@ -128,12 +134,12 @@ public class AdminEventController {
     @GetMapping("/delete/{eventId}")
     public String deleteEvent(@PathVariable Integer eventId){
         Event event = eventRepository.findByEventId(eventId);
-        eventRepository.delete(event);
-        return "redirect:/admin/event/list/0";
+        if(event != null){
+            eventLocationRepository.deleteEventLocationsByEvent_EventId(eventId);
+            eventRepository.delete(event);
+            return "redirect:/admin/event/list/0";
+        }else{
+            return "redirect:/admin/event/{eventId}";
+        }
     }
-
-
-
-    // TODO : Update 메소드 필요.
-    // TODO : Delete 메소드도 필요.
 }
